Refactor engine imports to enhance modularity and maintainability across game modules
