+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "time-pilot",
   "private": true,
-  "version": "28.1.0",
+  "version": "28.2.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
