+++ src/game/bullet.ts
@@ -1,8 +1,7 @@
 /* Converted from TimePilot.Bullet.js (AMD) to ESM TypeScript. */
 import userOptions from "./user-options";
 import { drawDebugVectors } from "./debug-vectors";
-import { Sound as SoundEngine } from "./engine";
-import helpers from "./engine/helpers";
+import { helpers, Sound as SoundEngine } from "./engine";
 import palette from "./palette";
 import { getDespawnRadius } from "./viewport";
 import type {
