+++ src/game/player.ts
@@ -2,8 +2,7 @@
 import { levels, player, scoring, sounds } from "./constants";
 import userOptions from "./user-options";
 import { drawDebugVectors } from "./debug-vectors";
-import { Sound as SoundEngine } from "./engine";
-import helpers from "./engine/helpers";
+import { helpers, Sound as SoundEngine } from "./engine";
 import palette from "./palette";
 import type {
   BulletFactoryInstance,
